+++ lib/app/ui/novel_dir_app.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:flutter/widgets.dart';
 import 'package:novel_v3/app/ui/routes_helper.dart';
 import 'package:novel_v3/app/core/models/novel_pdf.dart';
-import 'package:novel_v3/more_libs/pdf_readers_v1.2.3/pdf_reader.dart';
+import 'package:novel_v3/more_libs/pdf_readers_v1.3.3/pdf_reader.dart';
 import '../services/index.dart';
 
 export '../services/index.dart';