+++ lib/app/ui/routes_helper.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:novel_v3/app/others/chapter_reader/chapter_reader_config.dart';
 import 'package:novel_v3/app/others/chapter_reader/chapter_reader_screen.dart';
 import 'package:novel_v3/app/ui/main_ui/screens/content/novel_content_home_screen.dart';
-import 'package:novel_v3/more_libs/pdf_readers_v1.2.3/screens/pdfrx_reader_screen.dart';
-import 'package:novel_v3/more_libs/pdf_readers_v1.2.3/types/pdf_config.dart';
+import 'package:novel_v3/more_libs/pdf_readers_v1.3.3/screens/pdfrx_reader_screen.dart';
+import 'package:novel_v3/more_libs/pdf_readers_v1.3.3/types/pdf_config.dart';
 import 'package:novel_v3/more_libs/setting_v2.0.0/others/path_util.dart';
 import 'package:provider/provider.dart';