+++ lib/main.dart
@@ -7,7 +7,7 @@ import 'package:novel_v3/more_libs/fetcher_v1.0.0/fetcher.dart';
 import 'package:novel_v3/more_libs/general_static_server/constants.dart';
 import 'package:novel_v3/more_libs/general_static_server/general_server.dart';
 import 'package:novel_v3/more_libs/novel_v3_uploader_v1.3.0/constants.dart';
-import 'package:novel_v3/more_libs/pdf_readers_v1.2.3/pdf_reader.dart';
+import 'package:novel_v3/more_libs/pdf_readers_v1.3.3/pdf_reader.dart';
 import 'package:novel_v3/more_libs/setting_v2.0.0/setting.dart';
 import 'package:provider/provider.dart';
 import 'package:t_widgets/t_widgets.dart';